fixed migration

+++ database/migrations/2025_09_09_142531_create_group_rules_positions_table.php
@@ -15,7 +15,7 @@ class CreateGroupRulesPositionsTable extends Migration
     {
         Schema::create('group_rules_positions', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedBigInteger('group_rules_id');
+            $table->unsignedInteger('group_rules_id');
             $table->string('positions');
             $table->string('color');
             $table->string('label');
